Refactored and improved the codebase
