+++ .env.example
@@ -1,4 +1,9 @@
 # Dashboard configuration
 DASHBOARD_PORT=5000
 STATUS_FILES_PATH=/var/log/ansible/status
+RELEASE_FILES_PATH=/var/log/ansible/releases
+DATABASE_FILES_PATH=/var/log/ansible/data
 TIMEZONE=Europe/Amsterdam
+# Sysadm user and group
+UID=1000
+GID=1000
