+++ Dockerfile
@@ -9,6 +9,9 @@ COPY . .
 
 ENV FLASK_APP=app.py
 ENV FLASK_ENV=production
-ENV DATABASE_FILE=/data/dashboard.db
+
+RUN chmod +x entrypoint.sh
+
+ENTRYPOINT [ "./entrypoint.sh" ]
 
 CMD ["sh", "-c", "python watcher.py & flask run --host=0.0.0.0 --port=${FLASK_PORT}"]
