+++ docker-compose.yml
@@ -1,20 +1,30 @@
 services:
   dashboard:
     container_name: ansible-dashboard
+    uid: "${ UID }" 
+    gid: "${ GID }"
     build: .
     image: ansible-dashboard
     network_mode: host
+    logging:
+      driver: syslog
+      options:
+        syslog-address: "udp://logging.kdev:5515"
+        syslog-format: rfc5423
+        tag: "sysadm_status-dashboard"
+        syslog-facility: "daemon"
     # Not needed when using network_mode: host
     # ports:
     #   - "0.0.0.0:${DASHBOARD_PORT:-5000}:5000"
     volumes:
-      - ${STATUS_FILES_PATH:-/var/log/ansible/status}:/data/status
-      - ${RELEASE_FILES_PATH:-/var/log/ansible/status/releases}:/data/releases
-      - ./data/db:/data/db
+      # Mapping host filesystem to container filesystem
+      - ${STATUS_FILES_PATH}:/data/status
+      - ${RELEASE_FILES_PATH}:/data/releases
+      - ${DATABASE_FILES_PATH}:/data/db
     environment:
-      - FLASK_PORT=${DASHBOARD_PORT:-5000}
+      - FLASK_PORT=${DASHBOARD_PORT}
       - STATUS_DIR=/data/status
       - RELEASE_DIR=/data/releases
       - DATABASE_FILE=/data/db/dashboard.db
-      - TZ=${TIMEZONE:-Europe/Amsterdam}
+      - TZ=${TIMEZONE}
     restart: unless-stopped
