+++ templates/base.html
@@ -1,0 +1,204 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>System Status Dashboard</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.2/font/bootstrap-icons.css" rel="stylesheet">
+    <link href="https://cdn.datatables.net/1.13.7/css/dataTables.bootstrap5.min.css" rel="stylesheet">
+    <style>
+        .status-ok { background-color: #14e244; }
+        .status-failed { background-color: #f8d7da; }
+        .status-unknown { background-color: #fff3cd; }
+        .timestamp {
+            font-size: 0.8em;
+            color: #666;
+        }
+        .hosts {
+            font-size: 0.9em;
+            color: #444;
+        }
+        .details {
+            font-size: 0.9em;
+            color: #721c24;
+            margin-top: 5px;
+        }
+        .release-state {
+            display: flex;
+            gap: 1rem;
+        }
+        .state-indicator {
+            display: flex;
+            align-items: center;
+            gap: 0.5rem;
+            padding: 0.25rem 0.5rem;
+            border-radius: 4px;
+            background-color: #f8f9fa;
+        }
+        .state-indicator.active {
+            background-color: #e8f5e9;
+        }
+        .state-indicator i {
+            font-size: 1.2em;
+        }
+        .state-indicator.active i {
+            color: #198754;
+        }
+        .state-indicator.inactive i {
+            color: #dc3545;
+        }
+        .action-history {
+            max-height: 400px;
+            overflow-y: auto;
+        }
+        .action-row {
+            padding: 1rem;
+            border-bottom: 1px solid #eee;
+        }
+        .action-row:last-child {
+            border-bottom: none;
+        }
+        .action-details {
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
+            gap: 1rem;
+        }
+        .action-meta {
+            display: flex;
+            align-items: center;
+            gap: 1.5rem;
+        }
+        .git-tag {
+            color: #666;
+            white-space: nowrap;
+        }
+        .action-type {
+            display: inline-block;
+            padding: 0.25em 0.6em;
+            font-size: 0.85em;
+            font-weight: 600;
+            border-radius: 0.25rem;
+        }
+        .action-type.release {
+            background-color: #cff4fc;
+            color: #055160;
+        }
+        .action-type.promote {
+            background-color: #d1e7dd;
+            color: #0a3622;
+        }
+        .action-type.rollback {
+            background-color: #f8d7da;
+            color: #58151c;
+        }
+        .modal-title {
+            font-size: 1.1rem;
+            font-weight: 600;
+        }
+        /* DataTables Customization */
+        .dataTables_wrapper .dataTables_filter {
+            margin-bottom: 1rem;
+        }
+        .dataTables_wrapper .dataTables_length {
+            margin-bottom: 1rem;
+        }
+        table.dataTable > tbody > tr.child ul.dtr-details {
+            display: block;
+        }
+        /* Table responsiveness */
+        .table-container {
+            max-width: 100%;
+            margin: 0 auto;
+            padding: 0 1rem;
+        }
+        .table {
+            width: 100% !important;
+        }
+        .table td {
+            white-space: nowrap;
+            overflow: hidden;
+            text-overflow: ellipsis;
+        }
+        .release-state {
+            min-width: auto;
+            flex-wrap: wrap;
+            gap: 0.5rem;
+        }
+        .git-tag {
+            font-size: 0.9em;
+            color: #666;
+        }
+        .modal-lg {
+            max-width: 1000px;
+        }
+    </style>
+</head>
+<body>
+    <div class="container-fluid mt-4">
+        <h1>System Status Dashboard</h1>
+        
+        <ul class="nav nav-tabs mb-3" id="dashboardTabs" role="tablist">
+            <li class="nav-item" role="presentation">
+                <a class="nav-link {% if active_tab == 'playbooks' %}active{% endif %}" 
+                   href="{{ url_for('playbooks') }}">
+                    Playbook Status
+                </a>
+            </li>
+            <li class="nav-item" role="presentation">
+                <a class="nav-link {% if active_tab == 'releases' %}active{% endif %}"
+                   href="{{ url_for('releases') }}">
+                    Release Status
+                </a>
+            </li>
+            <li class="nav-item" role="presentation">
+                <a class="nav-link {% if active_tab == 'hosts' %}active{% endif %}"
+                   href="{{ url_for('hosts') }}">
+                    Monitored Hosts
+                </a>
+            </li>
+        </ul>
+        
+        <div class="tab-content" id="dashboardContent">
+            {% block content %}{% endblock %}
+        </div>
+    </div>
+    
+    <!-- Scripts -->
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
+    <script src="https://code.jquery.com/jquery-3.7.1.min.js"></script>
+    <script src="https://cdn.datatables.net/1.13.7/js/jquery.dataTables.min.js"></script>
+    <script src="https://cdn.datatables.net/1.13.7/js/dataTables.bootstrap5.min.js"></script>
+    <script>
+        $(document).ready(function() {
+            // Initialize Playbooks table
+            $('#playbooks-table').DataTable({
+                order: [[1, 'desc']], // Sort by Last Run by default
+                columnDefs: [
+                    { orderable: false, targets: 3 } // Disable sorting for Hosts column
+                ],
+                pageLength: 25,
+                scrollX: false,
+                autoWidth: false
+            });
+            
+            // Initialize Releases table
+            $('#releases-table').DataTable({
+                order: [[1, 'desc']], // Sort by Last Updated by default
+                columnDefs: [
+                    { orderable: true, targets: [0, 1] }, // Only allow sorting on Binary and Last Updated
+                    { orderable: false, targets: [2, 3, 4, 5] }
+                ],
+                pageLength: 25,
+                scrollX: false,
+                autoWidth: false
+            });
+            
+            // Handle Bootstrap tab changes
+            $('button[data-bs-toggle="tab"]').on('shown.bs.tab', function (e) {
+                // Adjust DataTables column widths when showing a tab
+                $.fn.dataTable.tables({ visible: true, api: true }).columns.adjust();
+            });
+        });
+    </script>
+</body>
+</html>
